moved the delegation-related routines that need to be compiled without thumb to a new file and updated the Makefile accordingly.
