+++ scamp/scamp.h
@@ -314,6 +314,9 @@ extern uint cmd_srom(sdp_msg_t *msg);
 // scamp-isr.c
 
 extern void vic_setup(void);
+
+// scamp-del.c
+
 extern void img_cp_exe(void);
 
 // scamp-app.c
